refactor: 공통으로 사용되는 스타일 전역으로 분리
[#114]
