+++ frontend/src/components/common/Button/Button.tsx
@@ -1,5 +1,6 @@
+import { borderRadius, pillStyle } from 'style';
+import styled from 'styled-components';
 import type { CSSProp } from 'styled-components';
-import styled, { css } from 'styled-components';
 
 import type { ButtonHTMLAttributes, MouseEventHandler } from 'react';
 
@@ -61,22 +62,5 @@ const S = {
     ${({ css }) => css};
   `,
 };
-
-export const pillStyle = css`
-  height: 36px;
-  padding-top: 0;
-  padding-bottom: 0;
-  line-height: 18px;
-  font-size: 16px;
-  border-radius: 21px;
-`;
-
-export const borderRadius = (direction: DirectionType) => css`
-  ${direction === 'all' && 'border-radius: 0;'}
-  ${direction === 'top' && 'border-top-left-radius: 0;'}
-  ${direction === 'top' && 'border-top-right-radius: 0;'}
-  ${direction === 'bottom' && 'border-bottom-left-radius: 0;'}
-  ${direction === 'bottom' && 'border-bottom-right-radius: 0;'}
-`;
 
 export default Button;
